Add initial project setup with Astro
This commit includes the introduction of Astro project setup files, including the configuration. The main change is the addition of the configuration file `astro.config.mjs`, where Tailwind is set up for Astro integration. It also adds `.gitignore` files, a `package.json`, and setup files for VSCode and IntelliJ IDEA. The project includes a favicon and a simple Astro page. It comes with a `package-lock.json` file which locks down the versions of each dependency ensuring that installs generate the exact same tree across machines.