Add new "SF-BN B203" product description and implement product details page
The new product file "b203.json" was added to the product description containing detailed descriptions, features and product aspects. In the "prod.astro" file, a product details page was implemented, which fetches data from the JSON file and presents it to the user. The changes also included the development of a "ProductTabBtn" component to improve code modularity and the user interface.